add a bunch more unit tests
