+++ _fixtures/channel-breakroom.js
@@ -1,0 +1,12 @@
+const breakroomChannel = {
+  id: 'TWEI7XP8B',
+  is_archived: false,
+  is_channel: true,
+  is_member: true,
+  is_mpim: false,
+  is_private: false,
+  name: 'breakroom',
+  name_normalized: 'The Breakroom'
+};
+
+module.exports = breakroomChannel;
